Title: 실수 입력받아 계산, Time: 166ms, Memory: 11MB, Status: Passed - CodeTree

+++ 240415/실수 입력받아 계산/calculation-by-inputting-a-real-number.java
@@ -4,8 +4,8 @@ public class Main {
         // 여기에 코드를 작성해주세요.
         Scanner sc = new Scanner(System.in);
 
-        double a = sc.nextInt();
-        double b = sc.nextInt();
+        double a = sc.nextDouble();
+        double b = sc.nextDouble();
 
         System.out.printf("%.2f", a+b);
